+++ src/pages/Register/index.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
-import { useRegisterMutation } from '@/api/hooks/useGetRegister';
+import { useRegisterMutation } from '@/api/hooks/useRegister';
 import KAKAO_LOGO from '@/assets/kakao_logo.svg';
 import { Button } from '@/components/common/Button';
 import { UnderlineTextField } from '@/components/common/Form/Input/UnderlineTextField';
@@ -38,6 +38,7 @@ export const RegisterPage = () => {
       }
     }
   };
+
   return (
     <Wrapper>
       <Logo src={KAKAO_LOGO} alt="카카오 CI" />
